netlify deploy fix attempt 8: attempt to rename maybe case sensitive files
