+++ examples/demo-app/src/cloud-providers/foursquare/fsq-provider.js
